Update Readme.md
"Commit!" - shouted he with a quizzical accent of a stereotypical Asian man. Like that wail was one of the mandatory Kiai he had to bellow in front of a judge panel as part of a step toward his next belt in Kyokushin Karate. But there *was* no judge panel, for it was the second year of a pandemic, and the only mastery rank that cry earned him was that of an utter, committed degenerate. Who still fucking sucked at JavaScript, mind you.